pyproject.toml
[project]
name = "tns-serv"
version = "0.1.0"
description = "Server for the Teacher Notifier System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[all]>=0.123.3",
    "fastapi-admin>=1.0.4",
    "flask>=3.1.2",
    "flask-admin>=2.0.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pydantic[email]>=2.12.5",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "fastapi-utils>=0.8.0",
    "sqladmin[full]>=0.22.0",
    "typing-inspect>=0.9.0",
    "firebase-admin>=7.1.0",
]
